+++ src/human/Gender.java
@@ -1,0 +1,5 @@
+package human;
+
+public enum Gender {
+    Male, Femala
+}
